Use CloudNativePG to deploy Postgres (#29)
Toward #23 
- replace bitnami postgresql helm chart with the CNPG operator and helm
chart. Requires the operator to first be installed (this can be done
with helm)
- three instances of Postgres are created by default (a primary and two
replicas)
- the operator requires the credentials to be stored in a secret (it's
not possible to pass them in plain text). The secret needs to be created
before installing the chart
- configure the `xnat-properties.conf` file in an init container rather
than as a secret. This is because we no longer know the password for
postgres (it's stored in a secret)
- backups are not yet configured but CNPG should handle them for us. I
can do this in a separate PR once we know where we'll store the backups
- update docs with changes
- update chart version to 0.0.8

---------

Co-authored-by: Daniel Matthews <d.matthews@ucl.ac.uk>